Fix possible crash on bad return address
The return address may be occasionally overwritten
due to a buffer overflow or another mistake, or even
intentionally e.g. consider the following code:

test() {
	new i = 0x12345678;
	#emit load.s.pri i
	#emit stor.s.pri 0
	#emit stor.s.pri 4
	#emit stor.s.pri 8
	#emit stor.s.pri 12
	new a[1];
	a[i] = 0;
}

Here we overwrite 4 stack cells right before the local
variable 'i', including return address, old frame and size of
parameters. This is a 100% crash on the old version.

The obvious solution is to check callAddress_ before dereferencing
it in AMXStackFrame ctor.