Move long call check to a function:
Use a copy of the AMX struct instead of a few preserved values.

Check long calls only on `BREAK` statements.  It's far less likely to conflict with other code by mangling the stack that way.

Move the long call checks inside the debug existence checks.

Use the top of the heap, not the bottom.